✨ feat(sessions): share session-schema + dirty tracking telemetry (B.1)
Extract normalizeChatRow/GroupRow/Scalars into src/state/session-schema.mjs for server+client; kill the ensureChatShape twin. Track dirtyChatIds/groups/scalars and deletedChatIds for upcoming PATCH, but still PUT the whole blob. Dev flush verifier warns on unmarked chat mutations.
