+++ documentation/context.md
@@ -82,9 +82,9 @@ Minnow/
 
 **Secrets:** AES-256-GCM at rest; key file `~/.minnow/.key` (`0o600`). Rotating or deleting `.key` makes encrypted secrets unrecoverable.
 
-**Canonical session store (HTTP):** `sessions/sessions.db` (SQLite) — GET/PUT `/api/config/sessions` still exchange the whole SessionState blob (SPA still PUT for now). Server `validateSessionState` / client `ensureChatShape` keep the full Chat allowlist (including todos, token ledger, pending queues, etc. via named passthrough on the server); there is no `MAX_CHATS` hard-trim on save. Kitchen-sink contract: [`test/fixtures/migration/kitchen-sink-sessions-state.json`](../test/fixtures/migration/kitchen-sink-sessions-state.json).
+**Canonical session store (HTTP):** `sessions/sessions.db` (SQLite) — GET/PUT `/api/config/sessions` still exchange the whole SessionState blob (SPA still PUT for now). Chat/group/scalar normalization is shared in [`src/state/session-schema.mjs`](../src/state/session-schema.mjs) (`normalizeChatRow` / `normalizeGroupRow` / `normalizeSessionScalars`) — imported by server validators and client `ensureChatShape` (thin wrapper; no client twin). There is no `MAX_CHATS` hard-trim on save. Kitchen-sink contract: [`test/fixtures/migration/kitchen-sink-sessions-state.json`](../test/fixtures/migration/kitchen-sink-sessions-state.json).
 
-**Sessions SQLite (Phase A + B.0):** On first open, legacy `sessions/state.json` is imported once and renamed to `state.json.migrated`. Whole-blob R/W lives in [`server/config/sessions-repo.js`](../server/config/sessions-repo.js); `terminalHistory` is server-owned on PUT/PATCH (only `appendTerminalRun` writes it). **PATCH `/api/config/sessions`** accepts `{ baseVersion, chats?, deleteChatIds?, groups?, deleteGroupIds?, scalars? }` — absent keys mean unchanged; deletes are explicit id lists; dirty chats/groups are full objects. Implemented via `patchResource` → `patchSessionState` (validators: `normalizeChatRow` / `normalizeGroupRow` / `normalizeSessionScalars`). Headless [`src/headless/persist-chat.ts`](../src/headless/persist-chat.ts) uses PATCH (no GET-splice-PUT). A rotating `state.json.backup` mirror flushes on a 5-minute dirty debounce and on `closeSessionsDb()`. Rollback: `MINNOW_SESSIONS_STORE=json`. Export: `POST /api/config/sessions/export-json`. Hot server consumers use indexed SQLite point lookups — see [`plans/sessions-sqlite-migration.md`](plans/sessions-sqlite-migration.md).
+**Sessions SQLite (Phase A + B.0 + B.1):** On first open, legacy `sessions/state.json` is imported once and renamed to `state.json.migrated`. Whole-blob R/W lives in [`server/config/sessions-repo.js`](../server/config/sessions-repo.js); `terminalHistory` is server-owned on PUT/PATCH (only `appendTerminalRun` writes it). **PATCH `/api/config/sessions`** accepts `{ baseVersion, chats?, deleteChatIds?, groups?, deleteGroupIds?, scalars? }` — absent keys mean unchanged; deletes are explicit id lists; dirty chats/groups are full objects. Implemented via `patchResource` → `patchSessionState`. Headless [`src/headless/persist-chat.ts`](../src/headless/persist-chat.ts) uses PATCH (no GET-splice-PUT). **B.1 dirty tracking (telemetry only):** SPA `touchChat` / `dirtyChatIds` / `deletedChatIds` / `dirtyGroupIds` / `sessionScalarsDirty` record what *would* PATCH; flush still PUTs the whole blob. Dev builds compare chats against a shadow copy at flush and `console.warn` unmarked mutations. A rotating `state.json.backup` mirror flushes on a 5-minute dirty debounce and on `closeSessionsDb()`. Rollback: `MINNOW_SESSIONS_STORE=json`. Export: `POST /api/config/sessions/export-json`. Hot server consumers use indexed SQLite point lookups — see [`plans/sessions-sqlite-migration.md`](plans/sessions-sqlite-migration.md).
 
 Full directory map: [`guides/configuration.md`](guides/configuration.md). Notable paths:
 
@@ -188,7 +188,7 @@ Built-in non-skill slash commands live in [`src/chat/slash-commands/registry.ts`
 | **`/goal`** | Work until a completion condition; post-turn evaluator continues the chat | `chat.activeGoal` ([`src/chat/goal/`](../src/chat/goal/)) |
 | **`/loop`** | Re-run a prompt on a fixed interval or self-paced delay while the app is open and the chat is idle | `chat.activeLoops[]` ([`src/chat/loop/`](../src/chat/loop/)) |
 
-**`/loop` modes:** `/loop 5m <prompt>` (interval; units `s`/`m`/`h`/`d`, sub-minute rounds up to 1m); `/loop <prompt>` (auto delay 1–60m from output change); bare `/loop` (maintenance: `<workspace>/.minnow/loop.md` or built-in checklist). Loops expire after 7 days. A global ticker in [`src/chat/loop/ticker.ts`](../src/chat/loop/ticker.ts) (started from [`src/main.ts`](../src/main.ts)) wakes at each loop's persisted `dueAt` (with a 15s safety poll) so reload/sleep survive and countdown matches fire time. Fires go through [`sendProgrammaticChatText`](../src/tools/loop.ts) so looped text gets full slash/skill resolution. `/goal` and `/loop` are mutually exclusive on a chat. `/clear` clears both. `activeLoops` persist in session storage (client `ensureChatShape` + server `validateSessionState`). Chat panel: [`src/ui/loop-status.ts`](../src/ui/loop-status.ts) (countdown, interval edit, pause/resume, skip, stop); skip marks the loop due immediately via [`src/chat/loop/skip.ts`](../src/chat/loop/skip.ts) and triggers an immediate ticker wake. Re-synced after transcript paint and OS app foreground changes. Sidebar rows with active loops show a masked [`/icons/loop.svg`](../public/icons/loop.svg) indicator ([`src/ui/chat-item-loop-icon.ts`](../src/ui/chat-item-loop-icon.ts)): rotates while at least one loop is unpaused, static when all are paused.
+**`/loop` modes:** `/loop 5m <prompt>` (interval; units `s`/`m`/`h`/`d`, sub-minute rounds up to 1m); `/loop <prompt>` (auto delay 1–60m from output change); bare `/loop` (maintenance: `<workspace>/.minnow/loop.md` or built-in checklist). Loops expire after 7 days. A global ticker in [`src/chat/loop/ticker.ts`](../src/chat/loop/ticker.ts) (started from [`src/main.ts`](../src/main.ts)) wakes at each loop's persisted `dueAt` (with a 15s safety poll) so reload/sleep survive and countdown matches fire time. Fires go through [`sendProgrammaticChatText`](../src/tools/loop.ts) so looped text gets full slash/skill resolution. `/goal` and `/loop` are mutually exclusive on a chat. `/clear` clears both. `activeLoops` persist in session storage (shared `normalizeChatRow` via client `ensureChatShape` + server `validateSessionState`). Chat panel: [`src/ui/loop-status.ts`](../src/ui/loop-status.ts) (countdown, interval edit, pause/resume, skip, stop); skip marks the loop due immediately via [`src/chat/loop/skip.ts`](../src/chat/loop/skip.ts) and triggers an immediate ticker wake. Re-synced after transcript paint and OS app foreground changes. Sidebar rows with active loops show a masked [`/icons/loop.svg`](../public/icons/loop.svg) indicator ([`src/ui/chat-item-loop-icon.ts`](../src/ui/chat-item-loop-icon.ts)): rotates while at least one loop is unpaused, static when all are paused.
 
 Naming: `src/tools/loop.ts` is the tool-call/send loop — unrelated to `/loop`. Do not rename it when touching session loops.
