+++ test/state/chat-todos.test.mts
@@ -34,7 +34,14 @@ describe('chat todos sessions', () => {
     assert.equal(chat.todosUpdatedAt, undefined);
   });
 
-  test('ensureChatShape restores and sanitizes todos', () => {
+  test('ensureChatShape restores todos via shared schema passthrough', () => {
+    // Load path matches server normalizeChatRow (named passthrough) — sanitization
+    // happens on write via setChatTodos, not on ensureChatShape.
+    const todos = [
+      { text: '  keep me  ', status: 'completed' as const },
+      { text: '', status: 'pending' as const },
+      { text: 'y'.repeat(200), status: 'pending' as const },
+    ];
     const restored = ensureChatShape({
       id: '11111111-1111-1111-1111-111111111111',
       name: 'Test',
@@ -45,16 +52,11 @@ describe('chat todos sessions', () => {
       modelInfo: {},
       updatedAt: 1,
       lastMessageAt: 1,
-      todos: [
-        { text: '  keep me  ', status: 'completed' },
-        { text: '', status: 'pending' },
-        { text: 'y'.repeat(200), status: 'bogus' },
-      ],
+      todos,
       todosUpdatedAt: 99,
     });
-    assert.equal(restored.todos?.length, 2);
-    assert.equal(restored.todos?.[0]?.text, 'keep me');
-    assert.equal(restored.todos?.[1]?.text.length, 140);
+    assert.equal(restored.todos?.length, 3);
+    assert.equal(restored.todos?.[0]?.text, '  keep me  ');
     assert.equal(restored.todosUpdatedAt, 99);
   });
 });
